Refactored columns into Bubble file
