+++ src/components/lab-notebook/DoubleColumn.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { ref, onMounted } from 'vue'
+import Bubble from './Bubble.vue';
 
 const title = ref()
 const id = ref("")
@@ -12,19 +13,20 @@ onMounted(() => {
 </script>
 
 <template>
-    <div class="flex flex-col bg-notebookBg p-8 rounded-[3.2em]">
-        <h2 :id="id" ref="title" class="font-title font-medium text-2xl mb-4">
+    <Bubble>
+        <template #title>
             <slot name="title"></slot>
-        </h2>
-        <div class="flex gap-8 w-full">
-            <div class="w-1/2 text-notebookText bg-slate p-6 rounded-[3.2em]">
-                <slot name="left">
-            </slot>
+        </template>
+        <template #body>
+            <div class="flex gap-8 w-full">
+                <div class="w-1/2 text-notebookText bg-slate p-6 rounded-[3.2em]">
+                    <slot name="left">
+                </slot>
+                </div>
+                <div class="w-1/2 text-notebookText bg-slate p-6 rounded-[3.2em]">
+                    <slot name="right"></slot>
+                </div>
             </div>
-            <div class="w-1/2 text-notebookText bg-slate p-6 rounded-[3.2em]">
-                <slot name="right"></slot>
-            </div>
-        </div>
-        
-    </div>
+        </template>
+    </Bubble>
 </template>
